Create view which adds an input string to the list
The `AddMovieActivity` contains a text field and a button, with
which the user can add a movie to the list in the `MainActivity`.

As of now, the movie list is cleared each time someone adds a
movie, i.e. it's recreated each time the `MainActivity` is
initialized. This behavior is of course unwanted, but later
we'll add movies to the database directly, which in turn will
be hooked up to the list in the `MainActivity`.