Make custom autocomplete adapter work with Movies.
We need custom objects (not just plain strings) in the auto complete dropdown, since we want to fetch the IMDb ID from the selected item and use in the view itself and later save to the database.

I noticed considerable performance loss when making AutoCompleteAdapter take objects of the type JSONObject from the API response. Tried a few performance tricks in the getView method, but after som test drive I concluded that there were some nasty lag when typing.

So for now, the adapter takes Movies.